Revert "Added Main window, fxml, css, and Controller class"
This reverts commit d27d191510d6d35feb6efa12d42c59656ba752a3.
