+++ Client/Style.css
@@ -1,0 +1,15 @@
+
+
+
+/*Put CSS stuff here plzzze*/
+
+.MainScreen {
+
+    /*-fx-background-image: url("/src/main/resources/cs4b/proj1/img/gameImage.jpg");*/
+    -fx-graphic: url('/src/main/resources/cs4b/proj1/img/gameImage.jpg');
+    -fx-background-repeat: stretch;
+    -fx-background-size: 500 400;
+    -fx-background-position: center center;
+    -fx-effect: dropshadow(three-pass-box, black, 30, 0.5, 0, 0);
+
+}
