+++ Client/sample.fxml
@@ -1,0 +1,74 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.geometry.Rectangle2D?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Menu?>
+<?import javafx.scene.control.MenuBar?>
+<?import javafx.scene.control.MenuItem?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.image.Image?>
+<?import javafx.scene.image.ImageView?>
+<?import javafx.scene.layout.Pane?>
+<?import javafx.scene.text.Font?>
+<?import javafx.scene.text.Text?>
+
+<Pane id="MainScreen" maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308" prefHeight="450.0" prefWidth="364.0" style="-fx-background-image: img\gameImage.jpg; -fx-background-color: skyblue;" stylesheets="@Style.css" xmlns="http://javafx.com/javafx/8.0.172-ea" xmlns:fx="http://javafx.com/fxml/1" fx:controller="cs4b.proj1.Controller">
+   <children>
+      <ImageView fitHeight="450.0" fitWidth="389.0" layoutX="1.0" layoutY="1.0" pickOnBounds="true" preserveRatio="true">
+         <image>
+            <Image url="@img/gameImage.jpg" />
+         </image>
+         <viewport>
+            <Rectangle2D />
+         </viewport>
+      </ImageView>
+      <Button fx:id="buttonOption2" layoutX="101.0" layoutY="211.0" mnemonicParsing="false" onMouseClicked="#button2Click" prefHeight="31.0" prefWidth="171.0" text="Against Computer - Easy" textFill="#2686e0">
+         <font>
+            <Font name="Impact" size="15.0" />
+         </font></Button>
+      <Button fx:id="buttonOption3" layoutX="101.0" layoutY="274.0" mnemonicParsing="false" onMouseClicked="#button3Click" prefHeight="31.0" prefWidth="171.0" text="Against Computer - Hard" textFill="#d11f1f">
+         <font>
+            <Font name="Impact" size="15.0" />
+         </font></Button>
+      <Button fx:id="buttonOption1" layoutX="101.0" layoutY="148.0" mnemonicParsing="false" onMouseClicked="#button1Click" prefHeight="31.0" prefWidth="171.0" text="Against Another Player" textFill="#e38120">
+         <font>
+            <Font name="Impact" size="15.0" />
+         </font></Button>
+      <MenuBar layoutY="-2.0" prefHeight="32.0" prefWidth="362.0" style="-fx-background-color: #3EA6B4;">
+        <menus>
+          <Menu mnemonicParsing="false" text="Info">
+            <items>
+              <MenuItem mnemonicParsing="false" text="About Our Game: &#10;Tic Tac Toe Game JavaFX implementation created by &#10;CS4B group members: Bobby, Keane, Naomi, Daniel. &#10;This is the first group project for our advance java class. &#10;Player is able to compete against another player, play &#10;against the computer in an easy level, or play against &#10;the computer in a hard level. Concepts of minimax &#10;search is implemented for a competitive AI. Player 1&#10;is 'X' and goes first, and player 2 is 'O'.&#10;" />
+            </items>
+          </Menu>
+        </menus>
+      </MenuBar>
+      <Text fill="#108be3" layoutX="155.0" layoutY="109.0" strokeType="OUTSIDE" strokeWidth="0.0" text="TAC">
+         <font>
+            <Font name="Impact" size="36.0" />
+         </font>
+      </Text>
+      <Text fill="#e57b19" layoutX="265.0" layoutY="109.0" strokeType="OUTSIDE" strokeWidth="0.0" text="TOE">
+         <font>
+            <Font name="Impact" size="36.0" />
+         </font>
+      </Text>
+      <Text fill="#e57b19" layoutX="45.0" layoutY="107.0" strokeType="OUTSIDE" strokeWidth="0.0" text="TIC">
+         <font>
+            <Font name="Impact" size="36.0" />
+         </font>
+      </Text>
+      <ImageView fitHeight="32.0" fitWidth="32.0" layoutX="110.0" layoutY="79.0" pickOnBounds="true" preserveRatio="true">
+         <image>
+            <Image url="@img/X.png" />
+         </image>
+      </ImageView>
+      <ImageView fitHeight="32.0" fitWidth="32.0" layoutX="222.0" layoutY="79.0" pickOnBounds="true" preserveRatio="true">
+         <image>
+            <Image url="@img/O.png" />
+         </image>
+      </ImageView>
+      <TextField fx:id="player1txt" layoutX="27.0" layoutY="353.0" prefHeight="31.0" prefWidth="119.0" promptText="Player 1 Name" />
+      <TextField fx:id="player2txt" layoutX="211.0" layoutY="353.0" prefHeight="31.0" prefWidth="119.0" promptText="Player 2 Name" />
+   </children>
+</Pane>
